make the constants service importable in a project
